fixup! Add helper scripts for enabling systemd user service on RHEL7

+++ server/pxf-service/src/scripts/enable-systemd-user-service.sh
@@ -1,12 +1,20 @@
 #!/usr/bin/env bash
 
-# TODO: move to PXF repo
+PXF_SBINDIR="$(cd -- "$(dirname -- "${BASH_SOURCE[0]}")" &>/dev/null && pwd)"
+PXF_HOME="$(cd -- "$(dirname -- "${PXF_SBINDIR}")" &>/dev/null && pwd)"
+
 if [[ "${EUID}" != 0 ]]; then
     echo "Please run as root (or via sudo)"
     exit 1
 fi
 
-# TODO: determine PXF_HOME instead of reading from env
-install -m 0644 "${PXF_HOME}/conf/user@.service" /usr/lib/systemd/system/
-systemctl enable "user@$(id -u gpadmin).service"
-systemctl start "user@$(id -u gpadmin).service"
+if [[ ! -e /usr/lib/systemd/system/user@.service ]]; then
+    echo "Installing ${PXF_HOME}/conf/user@.service into /usr/lib/systemd/system/"
+    install -m 0644 "${PXF_HOME}/conf/user@.service" /usr/lib/systemd/system/
+else
+    echo "There is already a user@service file in /usr/lib/systemd/system/; skipping installation..."
+fi
+
+echo "Enabling and starting user@$(id -ru gpadmin).service ..."
+systemctl enable "user@$(id -ru gpadmin).service"
+systemctl start "user@$(id -ru gpadmin).service"
